First attempt.
Docker container builds, but don't know if the rust build
works yet.

Centos6 and Ubuntu 12.04 don't have new enough C++ toolchains
to compile rust, so base this off ubuntu 14.04.

Likewise, no one seems to have a new enough node to install
the taskcluster-vcs caching git wrapper, so we include a
Makefile to download and install an upstream build.

This adds quite a bit to the image size, so given it doesn't
work it's probably not worthwhile to keep.