Use ENTRYPOINT instead of CMD.
Our parent container uses ENTRYPOINT, which CMD does not override.
That means we need to use it too to avoid running the parents
buildbot worker script instead of building rust like we want.

In both cases this can be overridden with the --entrypoint optoin:

  docker run -ti --entrypoint /bin/bash build-rust

Note that it must come before the image specification.