Use --target to build both stdlib targets.
Based on ruststrap's instructions, this should build both
32 and 64 bit standard libraries with the 64-bit-hosted
compiler. This avoids building the compiler itself for
both targets, considerably speeding up the build.

Add a disabled arm-linux-androideabi cross target as well.

Adding ${arm} to the target should allow building for
android and b2g, but requires that the build host have
an android ndk installed, so this is just a stub for now.