Port repack script to python.
We want to include cargo in our repacks, but the manifest linking
cargo to a particular rust release seems to only be available
in toml, which is hard to parse in shell.

An alternative would be to go back to using the channel-rust-stable
manifest and adding targets from channel-rustc-stable, but this
seems less brittle, and I believe this is the resource supporting rustup.