Add a function for repacking just cargo nightlies.
We want to use `cargo build --frozen` in automation to verify
our Cargo.toml doesn't reference anything requiring a network
fetch, but that's only available in nightly builds.

It seems there's no toml-style manifest for cargo builds,
so fall back to the list-of-files manifest and build something
similar to what we would have gotten from toml to pass to
the rest of the machinery. Then repack just the cargo package
for the requested host architecture.