Add a splat_rust.py script for updating gecko.
Rough script invoked with a path to a gecko source tree will
try to read a local tooltool `manifest.tt` file with rust
repacks and update the corresponding manifests in the gecko
source to match.

Note we expect the source `manifest.tt` to have version keys.
Right not that requires manual doctoring, but it still speeds
up the process immensely.